refactor: move tests to corresponding test class because of moved method
